pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = ""
authors = ["anton-akulenko <anton.htl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
loguru = ">=0.7.2"
python-dotenv = ">=1.0.1"
argparse = ">=1.4.0"
requests = "^2.31.0"
bs4 = "^0.0.2"
selenium = "^4.19.0"


[tool.poetry.group.dev.dependencies]
mypy = ">=1.9.0"
flake8 = "^7.0.0"
pep8-naming = "^0.13.3"
pytest = "^8.1.1"
black = "^24.4.0"
flake8-black = "^0.3.6"
ruff = "^0.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
